+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## Forma 1.0.0
 
-* Added `subParser` and `value` combinators. Changed signature of
-  `mkFieldError` to accept `NonEmpty (SelectedName names)` as first
-  argument.
+* Added `subParser` and `value` combinators.
+* Changed signature of `mkFieldError` to accept `NonEmpty (SelectedName names)` 
+  as first argument.
+* Added `runForm'` function that returns raw `BranchState`.
+* Exported `FieldError`'s value constructor.
 
 ## Forma 0.2.0